Updated deleteSection method
Updated deleteSection method to throw CONFLICT when trying to delete a section with current enrollments

+++ src/main/java/com/cst438/controller/SectionController.java
@@ -27,6 +27,9 @@ public class SectionController {
     UserRepository userRepository;
 
 
+    /****************************
+         CREATE SECTION
+     ****************************/
     // ADMIN function to create a new section
     @PostMapping("/sections")
     public SectionDTO addSection(@RequestBody SectionDTO section) {
@@ -76,13 +79,16 @@ public class SectionController {
         );
     }
 
+    /****************************
+          UPDATE SECTION
+     ****************************/
     // ADMIN function to update a section
     @PutMapping("/sections")
     public void updateSection(@RequestBody SectionDTO section) {
         // can only change instructor email, sec_id, building, room, times, start, end dates
         Section s = sectionRepository.findById(section.secNo()).orElse(null);
         if (s==null) {
-            throw  new ResponseStatusException( HttpStatus.NOT_FOUND, "section not found "+section.secNo());
+            throw new ResponseStatusException( HttpStatus.NOT_FOUND, "section not found "+section.secNo());
         }
         s.setSecId(section.secId());
         s.setBuilding(section.building());
@@ -102,20 +108,31 @@ public class SectionController {
         sectionRepository.save(s);
     }
 
-    // ADMIN function to create a delete section
-    // delete will fail if there are related assignments or enrollments
+    /****************************
+        DELETE SECTION
+     ****************************/
+    // ADMIN function to delete a section
+    //  delete will fail if there are related assignments or enrollments
     @DeleteMapping("/sections/{sectionno}")
     public void deleteSection(@PathVariable int sectionno) {
         Section s = sectionRepository.findById(sectionno).orElse(null);
+
+        if (!(s.getEnrollments().isEmpty())) {
+          throw new ResponseStatusException( HttpStatus.CONFLICT,
+            "Cannot delete a section with current enrollments.");
+        }
+
         if (s != null) {
             sectionRepository.delete(s);
         }
     }
 
-
+    /****************************
+          LIST SECTIONS
+     ****************************/
     // get Sections for a course with request params year, semester
-    // example URL   /course/cst363/sections?year=2024&semester=Spring
-    // also specify partial courseId   /course/cst/sections?year=2024&semester=Spring
+    //  example URL   /course/cst363/sections?year=2024&semester=Spring
+    //  also specify partial courseId   /course/cst/sections?year=2024&semester=Spring
     @GetMapping("/courses/{courseId}/sections")
     public List<SectionDTO> getSections(
             @PathVariable("courseId") String courseId,
@@ -150,8 +167,11 @@ public class SectionController {
         return dto_list;
     }
 
+    /****************************
+        GET INSTRUCTOR SECTIONS
+     ****************************/
     // get Sections for an instructor
-    // example URL  /sections?instructorEmail=dwisneski@csumb.edu&year=2024&semester=Spring
+    //  example URL  /sections?instructorEmail=dwisneski@csumb.edu&year=2024&semester=Spring
     @GetMapping("/sections")
     public List<SectionDTO> getSectionsForInstructor(
             @RequestParam("email") String instructorEmail,
@@ -185,6 +205,9 @@ public class SectionController {
         return dto_list;
     }
 
+    /****************************
+     LIST SECTIONS FOR ENROLLMENT
+     ****************************/
     // List available Sections for Enrollment
     @GetMapping("/sections/open")
     public List<SectionDTO> getOpenSectionsForEnrollment() {
